feat(header): display header according to role
