+++ WEB/src/components/headers/AdminHeader.vue
@@ -1,0 +1,117 @@
+<script setup lang="ts">
+  import logo from '@/assets/logo/logo.svg'
+    import { useUserStore } from '@/stores/user';
+    const userStore = useUserStore()
+  const showTab = () => {
+    document.getElementsByClassName('rightTab')[0].style.display = 'inline'
+  }
+  const hideTab = () => {
+    document.getElementsByClassName('rightTab')[0].style.display = 'none'
+  }
+</script>
+
+<template>
+  <div class="main">
+    <router-link
+      class="logo"
+      to="/"
+    >
+      <div class="logo_img">
+        <img
+          class="main_logo"
+          :src="logo"
+        />
+      </div>
+      <p class="logo_title">TaalToolBox</p>
+    </router-link>
+    <div class="tabs">
+      <router-link
+        class="tabs_element"
+        to="/"
+        >Acceuil</router-link
+      >
+      <router-link
+        class="tabs_element"
+        to="/map"
+        >Carte</router-link
+      >
+      <router-link
+        to="/add-card"
+        class="tabs_element"
+        >Ajouter</router-link
+      >
+      <router-link
+        to="/add-card"
+        class="tabs_element"
+        >Ajouter Prof</router-link
+      >
+      <router-link
+        to="/add-card"
+        class="tabs_element"
+        >Ajouter Elève</router-link
+      >
+      <router-link
+        to="/"
+        class="tabs_element"
+        >Classe</router-link
+      >
+    </div>
+
+    <router-link
+      class="conreg"
+      to="/login"
+      >Se connecter</router-link
+    >
+    <div
+      class="hamburger"
+      @click="showTab"
+    >
+      <!-- On construit le hamburger -->
+      <div class="build"></div>
+      <div class="build"></div>
+      <div class="build"></div>
+    </div>
+    <div class="rightTab">
+      <div class="content">
+        <div
+          class="hamburger_deploid"
+          @click="hideTab"
+        >
+          <!-- On construit le hamburger -->
+          <div class="build_deploid"></div>
+          <div class="build_deploid"></div>
+          <div class="build_deploid"></div>
+        </div>
+        <router-link
+          class="right_tab_element"
+          to="/"
+          >Acceuil</router-link
+        >
+        <router-link
+          class="right_tab_element"
+          to="/"
+          >Carte</router-link
+        >
+        <router-link
+          to="/add-card"
+          class="right_tab_element"
+          >Ajouter</router-link
+        >
+        <router-link
+          to="/"
+          class="right_tab_element"
+          >Classe</router-link
+        >
+        <router-link
+          to="/"
+          style="margin-top: 60vh; align-self: center"
+          >Se connecter</router-link
+        >
+      </div>
+    </div>
+  </div>
+</template>
+
+<style scoped>
+  @import './header.css';
+</style>
