+++ WEB/src/views/HomeView.vue
@@ -1,6 +1,9 @@
 <script setup lang="ts">
   import MainHeader from '@/components/headers/MainHeader.vue'
   import Greetings from '@/components/Greetings.vue'
+  import { useUserStore } from '@/stores/user';
+
+  const userStore = useUserStore()
 </script>
 
 <template>
@@ -12,6 +15,7 @@
 
 <style scoped>
   .page {
+    /* border : solid red 1px ;  */
     display: flex;
     flex-direction: column;
     align-items: center;
