+++ WEB/src/views/add-card/choiceHowToAddCardView.vue
@@ -1,6 +1,9 @@
 <script setup lang="ts">
   import MainHeader from '@/components/headers/MainHeader.vue'
   import choiceHowToAddCard from '@/components/card/add/choiceHowToAddCard.vue'
+  import { useUserStore } from '@/stores/user';
+
+  const useStore = useUserStore()
 </script>
 
 <template>
